+++ src/components/form/form_error_text/form_error_text.test.tsx
@@ -1,18 +1,15 @@
 import React from 'react';
 import { render } from 'enzyme';
-import { requiredProps } from '../../../test/required_props';
+import { requiredProps } from '../../../test';
 
 import { EuiFormErrorText } from './form_error_text';
 
 describe('EuiFormErrorText', () => {
   test('is rendered', () => {
     const component = render(
-      <EuiFormErrorText {...requiredProps}>
-        This is an error.
-      </EuiFormErrorText>
+      <EuiFormErrorText {...requiredProps}>This is an error.</EuiFormErrorText>
     );
 
-    expect(component)
-      .toMatchSnapshot();
+    expect(component).toMatchSnapshot();
   });
 });