+++ src/components/form/form_error_text/index.js
@@ -1,3 +1,0 @@
-export {
-  EuiFormErrorText,
-} from './form_error_text';